refactor: Kebab case EventTimer
