+++ src/event-timer.ts
